Change matrices to double precision; add error catches for bad inputs

+++ mmm/mmm.h
@@ -4,7 +4,7 @@
 // globals (anything here would be shared with all threads) */
 // I would declare the pointers to the matrices here (i.e., extern double **A, **B, **C),
 // as well as the size of the matrices, etc.
-extern int **mat1, **mat2, **mat3, **mat4;
+extern double **mat1, **mat2, **mat3, **mat4;
 extern int matdim;
 
 /** struct to hold parameters to mmm_par() **/
@@ -16,12 +16,12 @@ typedef struct thread_args
 } thread_args;
 
 void mmm_init();
-void mmm_reset(int **);
+void mmm_reset(double **);
 void mmm_freeup();
 void mmm_seq();
 void *mmm_par(void *args);
 double mmm_verify();
 void mmm_print();
-void mmm_print1(int **array);
+void mmm_print1(double **array);
 
 #endif /* MMM_H_ */
